feat: rename speakers avatar
